Fix crash scenarios.
Scenario 1: A very large number of files are processed, causing `identify` to send too much data to `stdout`. Fix: Perform `identify` calls one-by-one; this might be slower but it's definitely stabler.

Scenario 2: A large number of commands are passed to `convert`, so it executes ahead of time. It is possible to run exactly out of commands in this phase, leaving the final step to execute zero commands - which it will still try to do. Fix: skip the final step if there are zero commands remaining.

Scenario 3: A large quantity of data is loaded into memory. On low-resource machines (such as the Kubernetes pods on which gamefroot-texture-packer is intended to run), this can run out of memory, which may kill the machine. Fix: limit memory usage to 512MiB, forcing ImageMagick to fall back to disk cache. This doesn't appear to slow down any of the operations for which the packer is intended.