Fix more crash scenarios on cloud pods.
This packer is used on Kubernetes pods to process large batches of images. Sometimes this results in high memory usage, which triggers pod eviction, and of course the packer fails.

I've optimised the texture packing loop to avoid this kind of crash.

* Tasks are performed in series, not in parallel, so we can control how many ImageMagick instances are competing for memory at once.
* Detection for compacting multiple entries of the same file in `lib/generator.js:getImagesSizes()` is now less prone to throwing out all the images if it _doesn't_ find a [ character.
* `convert` command buffer is larger: about half the 128KiB limit of Linux `stdin` buffers.
* Explicit logging around `convert` `stdout`/`stderr` output.
* 512MiB memory limit on each `convert` command.
* Don't run a final command pass if all the commands have been drained.

This makes it much more stable under the conditions of a low-memory Kubernetes pod.